# Backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseApprovalRead": {
        "properties": {
          "approved_at": {
            "format": "date-time",
            "title": "Approved At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "recommendation_status": {
            "title": "Recommendation Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "video_delete_after": {
            "format": "date-time",
            "title": "Video Delete After",
            "type": "string"
          },
          "video_retention_days": {
            "title": "Video Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "recommendation_status",
          "approved_at",
          "video_delete_after",
          "video_retention_days"
        ],
        "title": "CaseApprovalRead",
        "type": "object"
      },
      "CaseApprovalRequest": {
        "properties": {
          "confirmed_route_types": {
            "items": {
              "$ref": "#/components/schemas/RouteType"
            },
            "title": "Confirmed Route Types",
            "type": "array"
          }
        },
        "required": [
          "confirmed_route_types"
        ],
        "title": "CaseApprovalRequest",
        "type": "object"
      },
      "CaseRead": {
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "classification": {
            "additionalProperties": true,
            "title": "Classification",
            "type": "object"
          },
          "critical_inconsistencies": {
            "title": "Critical Inconsistencies",
            "type": "integer"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactRead"
            },
            "title": "Facts",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "memory_image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryImageRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation_status": {
            "title": "Recommendation Status",
            "type": "string"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteRead"
            },
            "title": "Routes",
            "type": "array"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegment"
            },
            "title": "Segments",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRead"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventRead"
            },
            "title": "Timeline",
            "type": "array"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "video_stream_url": {
            "title": "Video Stream Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "analysis_id",
          "video_id",
          "status",
          "recommendation_status",
          "approved_at",
          "video_stream_url",
          "segments",
          "timeline",
          "facts",
          "classification",
          "sources",
          "routes",
          "critical_inconsistencies"
        ],
        "title": "CaseRead",
        "type": "object"
      },
...
      "FactReviewRequest": {
        "properties": {
          "action": {
            "enum": [
              "confirm",
              "correct"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "default": "",
            "maxLength": 1000,
            "title": "Reason",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "action"
        ],
        "title": "FactReviewRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TombstoneRead": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_role": {
            "title": "Actor Role",
            "type": "string"
          },
          "case_id_hash": {
            "title": "Case Id Hash",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "title": "Deleted At",
            "type": "string"
          }
        },
        "required": [
          "case_id_hash",
          "deleted_at",
          "action",
          "actor_role"
        ],
        "title": "TombstoneRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "SENDA API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit/tombstones/{case_id}": {
      "get": {
        "operationId": "read_tombstone_api_v1_audit_tombstones__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TombstoneRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Tombstone",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}": {
      "delete": {
        "operationId": "delete_case_api_v1_cases__case_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Case",
        "tags": [
          "cases"
        ]
      },
      "get": {
        "operationId": "read_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/approve": {
      "post": {
        "operationId": "approve_case_api_v1_cases__case_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseApprovalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/facts/{fact_id}": {
      "patch": {
        "operationId": "review_fact_api_v1_cases__case_id__facts__fact_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fact_id",
            "required": true,
            "schema": {
              "title": "Fact Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Review Fact",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/memory-image/content": {
      "get": {
        "operationId": "read_memory_image_content_api_v1_cases__case_id__memory_image_content_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Memory Image Content",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/memory-image/decision": {
      "post": {
        "operationId": "decide_memory_image_api_v1_cases__case_id__memory_image_decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryImageDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryImageDecisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Decide Memory Image",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/memory-image/regenerate": {
      "post": {
        "operationId": "regenerate_memory_image_api_v1_cases__case_id__memory_image_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryImageDecisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Regenerate Memory Image",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/rendered-video/retry": {
      "post": {
        "operationId": "retry_rendered_video_api_v1_cases__case_id__rendered_video_retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryImageDecisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Retry Rendered Video",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/rendered-video/stream": {
      "get": {
        "operationId": "stream_rendered_video_api_v1_cases__case_id__rendered_video_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Stream Rendered Video",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/routes/{route_id}/steps/{step_index}/narration": {
      "get": {
        "operationId": "narrate_route_step_api_v1_cases__case_id__routes__route_id__steps__step_index__narration_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_index",
            "required": true,
            "schema": {
              "title": "Step Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Narrate Route Step",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}